+++ crates/transaction-pool/src/best.rs
@@ -8,18 +8,12 @@ use reth_transaction_pool::{
     BestTransactions, CoinbaseTipOrdering, Priority, ValidPoolTransaction,
     error::InvalidPoolTransactionError, pool::BestTransactions as BestProtocolTransactions,
 };
-use std::{
-    sync::{
-        Arc,
-        mpsc::{self, Receiver, Sender},
-    },
-    thread::{self},
-};
+use std::sync::Arc;
 use tempo_evm::TempoTxResult;
 use tempo_precompiles::tip20::is_tip20_prefix;
 
 type TxOrdering = CoinbaseTipOrdering<TempoPooledTransaction>;
-type BestTransaction = Arc<ValidPoolTransaction<TempoPooledTransaction>>;
+pub type BestTransaction = Arc<ValidPoolTransaction<TempoPooledTransaction>>;
 type BestTransactionWithPriority = (BestTransaction, Priority<u128>);
 
 /// A best-transaction iterator that merges the protocol pool and the 2D nonces pool,
@@ -94,19 +88,10 @@ impl Iterator for MergeBestTransactions {
 
 impl BestTransactions for MergeBestTransactions {
     fn mark_invalid(&mut self, transaction: &Self::Item, kind: &InvalidPoolTransactionError) {
-        let next = if transaction.transaction.is_aa_2d() {
+        if transaction.transaction.is_aa_2d() {
             self.aa_2d_pool.mark_invalid(transaction, kind);
-            &mut self.next_aa_2d_pool
         } else {
             self.protocol_pool.mark_invalid(transaction, kind);
-            &mut self.next_protocol_pool
-        };
-
-        if next
-            .as_ref()
-            .is_some_and(|(tx, _)| is_invalidated_buffered_transaction(transaction, tx))
-        {
-            *next = None;
         }
     }
 
@@ -211,176 +196,6 @@ where
     fn set_skip_blobs(&mut self, skip_blobs: bool) {
         self.inner.set_skip_blobs(skip_blobs);
     }
-}
-
-/// Event returned by [`BestTransactionsStream`].
-#[derive(Debug)]
-pub enum BestTransactionsStreamEvent {
-    /// A transaction is ready for sequential payload execution.
-    Transaction(BestTransaction),
-    /// No transaction is currently buffered.
-    Empty,
-}
-
-enum BestTransactionsStreamCommand {
-    Invalid(InvalidTransaction),
-    NoUpdates,
-    SkipBlobs(bool),
-    Stop,
-}
-
-struct InvalidTransaction {
-    tx: BestTransaction,
-    kind: InvalidPoolTransactionError,
-    old_receiver: Receiver<BestTransactionsStreamEvent>,
-    new_sender: Sender<BestTransactionsStreamEvent>,
-}
-
-/// Drains a [`BestTransactions`] iterator into a channel while preserving delayed invalidation.
-pub struct BestTransactionsPrewarming {
-    rx: Receiver<BestTransactionsStreamEvent>,
-    commands_tx: Sender<BestTransactionsStreamCommand>,
-}
-
-impl BestTransactionsPrewarming {
-    /// Spawns a payload-scoped coordinator for `best_txs`.
-    pub fn new<Txs>(best_txs: Txs) -> Self
-    where
-        Txs: BestTransactions<Item = BestTransaction> + Send + 'static,
-    {
-        let (tx, rx) = mpsc::channel();
-        let (commands_tx, commands_rx) = mpsc::channel();
-        thread::spawn(move || {
-            run_best_transactions_coordinator(best_txs, tx, commands_rx);
-        });
-
-        Self { rx, commands_tx }
-    }
-}
-
-impl Drop for BestTransactionsPrewarming {
-    fn drop(&mut self) {
-        let _ = self.commands_tx.send(BestTransactionsStreamCommand::Stop);
-    }
-}
-
-impl Iterator for BestTransactionsPrewarming {
-    type Item = BestTransaction;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self.rx.recv() {
-            Ok(BestTransactionsStreamEvent::Transaction(tx)) => Some(tx),
-            Ok(BestTransactionsStreamEvent::Empty) | Err(_) => None,
-        }
-    }
-}
-
-impl BestTransactions for BestTransactionsPrewarming {
-    fn mark_invalid(&mut self, transaction: &Self::Item, kind: &InvalidPoolTransactionError) {
-        let (new_sender, new_receiver) = mpsc::channel();
-        let old_receiver = std::mem::replace(&mut self.rx, new_receiver);
-        let _ = self
-            .commands_tx
-            .send(BestTransactionsStreamCommand::Invalid(InvalidTransaction {
-                tx: transaction.clone(),
-                kind: kind.clone(),
-                old_receiver,
-                new_sender,
-            }));
-    }
-
-    fn no_updates(&mut self) {
-        let _ = self
-            .commands_tx
-            .send(BestTransactionsStreamCommand::NoUpdates);
-    }
-
-    fn set_skip_blobs(&mut self, skip_blobs: bool) {
-        let _ = self
-            .commands_tx
-            .send(BestTransactionsStreamCommand::SkipBlobs(skip_blobs));
-    }
-}
-
-fn run_best_transactions_coordinator<Txs>(
-    mut best_txs: Txs,
-    mut sender: Sender<BestTransactionsStreamEvent>,
-    command_receiver: Receiver<BestTransactionsStreamCommand>,
-) where
-    Txs: BestTransactions<Item = BestTransaction>,
-{
-    loop {
-        if let Some(tx) = best_txs.next() {
-            if sender
-                .send(BestTransactionsStreamEvent::Transaction(tx))
-                .is_err()
-            {
-                break;
-            }
-        } else {
-            if sender.send(BestTransactionsStreamEvent::Empty).is_err() {
-                break;
-            }
-        }
-
-        while let Ok(command) = command_receiver.try_recv() {
-            match command {
-                BestTransactionsStreamCommand::Invalid(invalid) => {
-                    process_invalid_transaction(&mut best_txs, invalid, &mut sender);
-                }
-                BestTransactionsStreamCommand::NoUpdates => best_txs.no_updates(),
-                BestTransactionsStreamCommand::SkipBlobs(skip_blobs) => {
-                    best_txs.set_skip_blobs(skip_blobs)
-                }
-                BestTransactionsStreamCommand::Stop => break,
-            }
-        }
-    }
-}
-
-fn process_invalid_transaction<Txs>(
-    best_txs: &mut Txs,
-    invalid: InvalidTransaction,
-    sender: &mut Sender<BestTransactionsStreamEvent>,
-) where
-    Txs: BestTransactions<Item = BestTransaction>,
-{
-    best_txs.mark_invalid(&invalid.tx, &invalid.kind);
-
-    while let Ok(event) = invalid.old_receiver.try_recv() {
-        if let BestTransactionsStreamEvent::Transaction(tx) = &event
-            && !is_invalidated_buffered_transaction(&invalid.tx, tx)
-        {
-            let _ = invalid.new_sender.send(event);
-        }
-    }
-
-    *sender = invalid.new_sender;
-}
-
-fn is_invalidated_buffered_transaction(
-    invalid: &BestTransaction,
-    candidate: &BestTransaction,
-) -> bool {
-    if invalid.transaction.is_expiring_nonce() {
-        return false;
-    }
-
-    if invalid.transaction.is_aa_2d() {
-        let Some(invalid_id) = invalid.transaction.aa_transaction_id() else {
-            return false;
-        };
-        return candidate
-            .transaction
-            .aa_transaction_id()
-            .is_some_and(|candidate_id| {
-                candidate_id.seq_id == invalid_id.seq_id && candidate_id.nonce >= invalid_id.nonce
-            });
-    }
-
-    !candidate.transaction.is_aa_2d()
-        && candidate.sender() == invalid.sender()
-        && candidate.nonce() >= invalid.nonce()
 }
 
 #[cfg(test)]
