+++ crates/transaction-pool/src/lib.rs
@@ -21,9 +21,7 @@ pub mod metrics;
 pub mod paused;
 pub mod tt_2d_pool;
 
-pub use best::{
-    BestTransactionsPrewarming, BestTransactionsStreamEvent, StateAwareBestTransactions,
-};
+pub use best::StateAwareBestTransactions;
 pub use maintain::TempoPoolUpdates;
 
 pub use metrics::{AA2dPoolMetrics, TempoPoolMaintenanceMetrics};
